Fix: Minor bug fixes

+++ packages/nc-gui/components/virtual-cell/OneToOne.vue
@@ -25,6 +25,10 @@ const clientMousePosition = inject(ClientMousePositionInj)
 
 const isExpandedFormOpen = inject(IsExpandedFormOpenInj, ref(false))
 
+const canvasCellEventData = inject(CanvasCellEventDataInj, reactive<CanvasCellEventDataInjType>({}))
+
+const cellEventHook = inject(CellEventHookInj, null)
+
 const cellClickHook = inject(CellClickHookInj, null)
 
 const onDivDataCellEventHook = inject(OnDivDataCellEventHookInj, null)
@@ -103,13 +107,30 @@ function onCellClick(e: Event) {
   }
 }
 
+const onCellEvent = (event?: Event) => {
+  if (!(event instanceof KeyboardEvent) || !event.target || isActiveInputElementExist(event) || !hasEditPermission.value) return
+
+  if (isExpandCellKey(event)) {
+    if (listItemsDlg.value) {
+      listItemsDlg.value = false
+    } else {
+      listItemsDlg.value = true
+    }
+
+    return true
+  }
+}
+
 onMounted(() => {
   onDivDataCellEventHook?.on(onCellClick)
   cellClickHook?.on(onCellClick)
+  cellEventHook?.on(onCellEvent)
 
   if (!hasEditPermission.value || !isCanvasInjected || !clientMousePosition || isExpandedFormOpen.value) return
 
   forcedNextTick(() => {
+    if (onCellEvent(canvasCellEventData.event)) return
+
     if (getElementAtMouse('.unlink-icon', clientMousePosition)) {
       unlinkRef(value.value)
     } else if (getElementAtMouse('.nc-canvas-table-editable-cell-wrapper .nc-plus.nc-action-icon', clientMousePosition)) {
@@ -123,6 +144,7 @@ onMounted(() => {
 onUnmounted(() => {
   onDivDataCellEventHook?.off(onCellClick)
   cellClickHook?.off(onCellClick)
+  cellEventHook?.off(onCellEvent)
 })
 </script>
